core: rename LiteSPICore to LiteSPIXIP
